+++ ng-semantic/input/input.d.ts
@@ -7,11 +7,15 @@ export declare class SemanticInputComponent {
 export declare class SemanticCheckboxComponent {
     control: FormControl;
     label: string;
+    disabled: boolean;
+    value: string | number;
+    name: string;
+    type: string;
+    private inputType;
+    private classType;
 }
 export declare class SemanticTextareaComponent {
     control: FormControl;
     label: string;
     rows: string;
-}
-export declare class SemanticFormComponent {
 }